Passive elements given in .npy file + History as attribute Topopt
Passive elements can be inputted through an .npy file. Can be used to optimize on a hexagonal base geometry.

History is now a attribute of the Topopt class. This allows continuing the optimization with new settings, allowing one to run the final iterations without a filter, this results in a more black/white design.